+++ src/utils/sendChannelNews.ts
@@ -1,4 +1,4 @@
-import { getNews, updateNews } from "database/news";
+import { addNews, getNews, updateNews } from "database/news";
 import { getSetting } from "database/settings";
 import {
   EmbedBuilder,
@@ -16,36 +16,48 @@ import { safeChannel, safeRole } from "./safeThings";
 /**
  * Sends news to a channel.
  * @param {Guild} guild Guild where the channel is in.
- * @param {string} id ID of the news.
  * @param {ChatInputCommandInteraction} interaction Command interaction.
- * @param {?string} title Title of the news.
- * @param {?string} body Content of the news.
+ * @param {{
+   title: string;
+   body: string;
+   author: string;
+   authorPFP: string;
+   id: string;
+   imageURL?: string | null;
+ }} newsOptions Options to send the news post.
+ * @param {?boolean} edit Whether or not should the function make a new message with some reused elements.
  * @returns News message in a channel.
  */
 export async function sendChannelNews(
   guild: Guild,
-  id: string,
   interaction: ChatInputCommandInteraction,
-  title?: string,
-  body?: string,
-  imageURL?: string,
-): Promise<void> {
-  const news = (await getNews(guild.id, id))!;
-  console.log(news.createdAt);
+  newsOptions: {
+    title: string;
+    body: string;
+    author: string;
+    authorPFP: string;
+    id: string;
+    imageURL?: string | null;
+  },
+  edit?: boolean,
+) {
+  const { title, body, author, authorPFP, imageURL, id } = newsOptions;
   const role = (await getSetting(guild.id, "news", "role")) as string;
   let roleToSend: Role | undefined;
   if (role) roleToSend = await safeRole(guild, role);
-  const avatar = news.authorPFP;
+
+  const news = (await getNews(guild.id, id))!;
+  const avatar = authorPFP;
   const embed = new EmbedBuilder()
     .setAuthor({
-      name: `${dotCheck({ string: avatar, doubleSpace: true })}${news.author}`,
+      name: `${dotCheck({ string: avatar, doubleSpace: true })}${author}`,
       iconURL: avatar,
     })
-    .setTitle(title ?? news.title)
-    .setDescription(body ?? news.body)
-    .setImage(imageURL ?? news.imageURL ?? null)
-    .setTimestamp(Number(news.updatedAt) || Number(news.createdAt))
-    .setFooter({ text: `Latest news from ${guild.name} • ID: ${news.id}` })
+    .setTitle(title)
+    .setDescription(body)
+    .setImage(edit ? (news.imageURL ?? null) : (imageURL ?? null))
+    .setTimestamp(edit ? news.createdAt : new Date())
+    .setFooter({ text: `Latest news from ${guild.name} • ID: ${id}` })
     .setColor(await colorize({ hue: Sokolors.Blue }));
 
   const channel = (await safeChannel(
@@ -63,12 +75,11 @@ export async function sendChannelNews(
   )
     return;
 
-  return await channel
-    .send({
-      embeds: [embed],
-      content: roleToSend ? mention(roleToSend.id, "ROLE") : undefined,
-    })
-    // Just noticed the call to updateNews, kinda weird but ok ?? This will set the updatedAt field in the DB tho
-    .then(async message => await updateNews(guild.id, id, undefined, undefined, message.id));
-    // Maybe just send the message and then create the DB entry if the message was sent successfully ?
+  const message = await channel.send({
+    embeds: [embed],
+    content: roleToSend ? mention(roleToSend.id, "ROLE") : undefined,
+  });
+
+  if (edit) return await updateNews(guild.id, id, title, body, message.id);
+  return await addNews(guild.id, title, body, author, authorPFP, message.id, imageURL, id);
 }